More DemoApp cleanup.  Cleaned up: -DynamicXYPlot -OrientationSensorExample -SimpleXYPlot
Also added other minor modifications such as removing unnecessary logging statements etc.

+++ AndroidPlot-Core/src/main/java/com/androidplot/util/PixelUtils.java
@@ -182,7 +182,7 @@ public class PixelUtils {
         if (matcher.matches()) {
             // -- Match found.
             // -- Extract value.
-            float value = Float.valueOf(matcher.group(1)).floatValue();
+            float value = Float.valueOf(matcher.group(1));
             // -- Extract dimension units.
             String unit = matcher.group(3).toLowerCase();
             // -- Get Android dimension constant.
